[FE] refactor: react query 폴더 병합
